move app shell for mfe out
